pyproject.toml
[tool.poetry]
name = "scigateway-auth"
version = "0.1.0"
description = "Authentication API for the SciGateway web application"
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/ral-facilities/scigateway-auth"
repository = "https://github.com/ral-facilities/scigateway-auth"
authors = [
    "Alan Kyffin <alan.kyffin@stfc.ac.uk>",
    "Alejandra Gonzalez Beltran <alejandra.gonzalez-beltran@stfc.ac.uk>",
    "Goel Biju",
    "Keiran Price <keiran.price@stfc.ac.uk>",
    "Louise Davies <louise.davies@stfc.ac.uk>",
    "Matthew Richards <matthew.richards@stfc.ac.uk>",
    "Patrick Austin <patrick.austin@stfc.ac.uk>",
    "Viktor Bozhinov <viktor.bozhinov@stfc.ac.uk>"
]

[tool.poetry.dependencies]
python = "^3.6"
Flask-RESTful = "0.3.7"
requests = "2.22.0"
PyJWT = "1.7.1"
cryptography = "2.8.0"
pyOpenSSL = "18.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
coverage = {extras = ["pyproject.toml"], version = "^5.4"}
pytest-cov = "^2.11.1"
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-broken-line = "^0.3.0"
flake8-bugbear = "^20.11.1"
flake8-builtins = "^1.5.3"
flake8-commas = "^2.0.0"
flake8-comprehensions = "^3.3.1"
flake8-import-order = "^0.18.1"
flake8-logging-format = "^0.6.0"
pep8-naming = "^0.11.1"
black = "^20.8b1"
safety = "^1.10.3"

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["scigateway_auth"]

[tool.coverage.run]
branch = true
source = ["scigateway_auth"]

[tool.coverage.report]
show_missing = true
